DockerServer/requirements.txt
fastapi>=0.70.0
uvicorn>=0.15.0
requests>=2.28.0
python-dotenv>=0.21.0
mido>=1.2.10
python-rtmidi>=1.4.9
